Move new examples to the root of the examples directory as running them from new directory fails in Playground as org.apache.beam.examples.basic namespace doesn't exist in the version of Beam deployed in the Playground.
