+++ examples/java/src/main/java/org/apache/beam/examples/ApproximateQuantilesExample.java
@@ -15,46 +15,49 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
+import java.util.List;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
+import org.apache.beam.sdk.transforms.ApproximateQuantiles;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.Min;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: MinPerKeyDemo
-//   description: Demonstration of Min.perKey transform usage.
+//   name: ApproximateQuantilesDemo
+//   description: Demonstration of ApproximateQuantiles transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 47
+//   context_line: 46
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
 //     - numbers
-//     - pairs
 
-public class MinPerKeyExample {
+public class ApproximateQuantilesExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
-    PCollection<KV<String, Integer>> input =
-        pipeline.apply(
-            Create.of(KV.of("a", 1), KV.of("a", 2), KV.of("b", 3), KV.of("b", 4), KV.of("b", 5)));
-    PCollection<KV<String, Integer>> minPerKey = input.apply(Min.perKey());
+    // Create a collection with numbers
+    PCollection<Integer> input = pipeline.apply(Create.of(1, 1, 2, 2, 3, 4, 4, 5, 5));
+    // This will produce a collection containing 5 values: the minimum value,
+    // Quartile 1 value, Quartile 2 value (median), Quartile 3 value, and the
+    // maximum value.
+    PCollection<List<Integer>> result = input.apply(ApproximateQuantiles.globally(5));
     // [END main_section]
-    // Log values
-    minPerKey.apply(ParDo.of(new LogOutput<>("PCollection numbers after Min transform: ")));
+    result.apply(
+        ParDo.of(
+            new LogOutput<>(
+                "PCollection numbers after ApproximateQuantiles.globally transform: ")));
     pipeline.run();
   }
