+++ examples/java/src/main/java/org/apache/beam/examples/CountPerKeyExample.java
@@ -15,14 +15,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
+import org.apache.beam.sdk.transforms.Count;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.GroupByKey;
 import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
@@ -30,35 +30,31 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: GroupByKeyDemo
-//   description: Demonstration of GroupByKey transform usage.
+//   name: CountPerKeyDemo
+//   description: Demonstration of Count.perKey transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 48
+//   context_line: 47
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - strings
+//     - numbers
 //     - pairs
-//     - group
 
-public class GroupByKeyExample {
+public class CountPerKeyExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
-    PCollection<KV<String, String>> pt =
+    PCollection<KV<String, Integer>> input =
         pipeline.apply(
-            Create.of(
-                KV.of("a", "apple"),
-                KV.of("a", "avocado"),
-                KV.of("b", "banana"),
-                KV.of("c", "cherry")));
-    PCollection<KV<String, Iterable<String>>> result = pt.apply(GroupByKey.create());
+            Create.of(KV.of("a", 1), KV.of("a", 2), KV.of("b", 3), KV.of("b", 4), KV.of("b", 5)));
+    PCollection<KV<String, Long>> countPerKey = input.apply(Count.perKey());
     // [END main_section]
-    result.apply(ParDo.of(new LogOutput<>("PCollection pairs after GroupByKey transform: ")));
+    // Log values
+    countPerKey.apply(ParDo.of(new LogOutput<>("PCollection numbers after Count transform: ")));
     pipeline.run();
   }
