+++ examples/java/src/main/java/org/apache/beam/examples/KvSwapExample.java
@@ -15,22 +15,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
-import org.apache.beam.sdk.transforms.Distinct;
 import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.KvSwap;
 import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: DistinctDemo
-//   description: Demonstration of Distinct transform usage.
+//   name: KvSwapDemo
+//   description: Demonstration of KvSwap transform usage.
 //   multifile: false
 //   default_example: false
 //   context_line: 46
@@ -39,23 +40,23 @@ import org.slf4j.LoggerFactory;
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - numbers
-//     - distinct
+//     - pairs
 
-public class DistinctExample {
+public class KvSwapExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
-
     // [START main_section]
-    // Create a collection with repeating numbers
-    PCollection<Integer> input = pipeline.apply(Create.of(1, 1, 2, 2, 3, 4, 4, 5, 5));
-    // Apply Distinct transform
-    PCollection<Integer> distinctNumbers = input.apply(Distinct.create());
+    // Create key/value pairs
+    PCollection<KV<String, Integer>> pairs =
+        pipeline.apply(
+            Create.of(KV.of("one", 1), KV.of("two", 2), KV.of("three", 3), KV.of("four", 4)));
+    // Returns KV collection with keys and values swapped: PCollection<KV<K,V>> ->
+    // PCollection<KV<V,K>>
+    PCollection<KV<Integer, String>> swap = pairs.apply(KvSwap.create());
     // [END main_section]
-    // Log values
-    distinctNumbers.apply(
-        ParDo.of(new LogOutput<>("PCollection numbers after Distinct.create transform: ")));
+    pairs.apply(ParDo.of(new LogOutput<>("PCollection element before KvSwap transform: ")));
+    swap.apply(ParDo.of(new LogOutput<>("PCollection element after KvSwap transform: ")));
     pipeline.run();
   }
