+++ examples/java/src/main/java/org/apache/beam/examples/MapElementsExample.java
@@ -15,63 +15,55 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.MapElements;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.transforms.join.CoGbkResult;
-import org.apache.beam.sdk.transforms.join.CoGroupByKey;
-import org.apache.beam.sdk.transforms.join.KeyedPCollectionTuple;
-import org.apache.beam.sdk.values.KV;
+import org.apache.beam.sdk.transforms.SimpleFunction;
 import org.apache.beam.sdk.values.PCollection;
-import org.apache.beam.sdk.values.TupleTag;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: CoGroupByKeyDemo
-//   description: Demonstration of CoGroupByKey transform usage.
+//   name: MapElementsDemo
+//   description: Demonstration of MapElements transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 54
+//   context_line: 47
 //   categories:
 //     - Core Transforms
-//     - Joins
 //   complexity: BASIC
 //   tags:
 //     - transforms
 //     - strings
-//     - integers
-//     - tuples
-//     - pairs
-//     - group
+//     - map
 
-public class CoGroupByKeyExample {
+public class MapElementsExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
-    // Create pairs
-    PCollection<KV<String, Integer>> pt1 =
-        pipeline.apply(Create.of(KV.of("a", 1), KV.of("b", 2), KV.of("b", 3), KV.of("c", 4)));
-    PCollection<KV<String, String>> pt2 =
-        pipeline.apply(
-            Create.of(
-                KV.of("a", "apple"),
-                KV.of("a", "avocado"),
-                KV.of("b", "banana"),
-                KV.of("c", "cherry")));
-
-    final TupleTag<Integer> t1 = new TupleTag<>();
-    final TupleTag<String> t2 = new TupleTag<>();
-    PCollection<KV<String, CoGbkResult>> result =
-        KeyedPCollectionTuple.of(t1, pt1).and(t2, pt2).apply(CoGroupByKey.create());
+    // Create collection of lowercase string
+    PCollection<String> strings = pipeline.apply(Create.of("one", "two", "three", "four"));
+    // Transform strings to upper case
+    PCollection<String> upperCaseStrings =
+        strings.apply(
+            MapElements.via(
+                new SimpleFunction<String, String>() {
+                  @Override
+                  public String apply(String s) {
+                    return s.toUpperCase();
+                  }
+                }));
     // [END main_section]
-    result.apply(ParDo.of(new LogOutput<>("PCollection pairs after CoGroupByKey transform: ")));
+    strings.apply(ParDo.of(new LogOutput<>("PCollection element before MapElements transform: ")));
+    upperCaseStrings.apply(
+        ParDo.of(new LogOutput<>("PCollection element after MapElements transform: ")));
     pipeline.run();
   }
