+++ examples/java/src/main/java/org/apache/beam/examples/MeanExample.java
@@ -15,22 +15,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.Mean;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.transforms.Sum;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: SumDemo
-//   description: Demonstration of Sum transform usage.
+//   name: MeanDemo
+//   description: Demonstration of Mean transform usage.
 //   multifile: false
 //   default_example: false
 //   context_line: 45
@@ -41,16 +41,16 @@ import org.slf4j.LoggerFactory;
 //     - transforms
 //     - numbers
 
-public class SumExample {
+public class MeanExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
     PCollection<Double> pc = pipeline.apply(Create.of(1.0, 2.0, 3.0, 4.0, 5.0));
-    PCollection<Double> sum = pc.apply(Sum.doublesGlobally());
+    PCollection<Double> mean = pc.apply(Mean.globally());
     // [END main_section]
     // Log values
-    sum.apply(ParDo.of(new LogOutput<>("PCollection numbers after Sum transform: ")));
+    mean.apply(ParDo.of(new LogOutput<>("PCollection numbers after Mean transform: ")));
     pipeline.run();
   }
