+++ examples/java/src/main/java/org/apache/beam/examples/MinPerKeyExample.java
@@ -15,66 +15,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
-import java.util.Arrays;
-import java.util.List;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.transforms.Count;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
+import org.apache.beam.sdk.transforms.Min;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.transforms.windowing.FixedWindows;
-import org.apache.beam.sdk.transforms.windowing.Window;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
-import org.joda.time.Duration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: WindowDemo
-//   description: Demonstration of Window transform usage.
+//   name: MinPerKeyDemo
+//   description: Demonstration of Min.perKey transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 54
+//   context_line: 47
 //   categories:
 //     - Core Transforms
-//     - Windowing
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - strings
-//     - timestamps
-//     - windows
+//     - numbers
+//     - pairs
 
-public class WindowExample {
+public class MinPerKeyExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
-
     // [START main_section]
-    // Create some input data with timestamps
-    List<String> inputData = Arrays.asList("foo", "bar", "foo", "foo");
-    List<Long> timestamps =
-        Arrays.asList(
-            Duration.standardSeconds(15).getMillis(),
-            Duration.standardSeconds(30).getMillis(),
-            Duration.standardSeconds(45).getMillis(),
-            Duration.standardSeconds(90).getMillis());
-
-    // Create a PCollection from the input data with timestamps
-    PCollection<String> items = pipeline.apply(Create.timestamped(inputData, timestamps));
-
-    // Create a windowed PCollection
-    PCollection<String> windowedItems =
-        items.apply(Window.into(FixedWindows.of(Duration.standardMinutes(1))));
-
-    PCollection<KV<String, Long>> windowedCounts = windowedItems.apply(Count.perElement());
+    PCollection<KV<String, Integer>> input =
+        pipeline.apply(
+            Create.of(KV.of("a", 1), KV.of("a", 2), KV.of("b", 3), KV.of("b", 4), KV.of("b", 5)));
+    PCollection<KV<String, Integer>> minPerKey = input.apply(Min.perKey());
     // [END main_section]
-    windowedCounts.apply(ParDo.of(new LogOutput<>("PCollection elements after Count transform: ")));
+    // Log values
+    minPerKey.apply(ParDo.of(new LogOutput<>("PCollection numbers after Min transform: ")));
     pipeline.run();
   }
