+++ examples/java/src/main/java/org/apache/beam/examples/PartitionExample.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
