+++ examples/java/src/main/java/org/apache/beam/examples/SumPerKeyExample.java
@@ -15,48 +15,46 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.KvSwap;
 import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.transforms.Sum;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: KvSwapDemo
-//   description: Demonstration of KvSwap transform usage.
+//   name: SumPerKeyDemo
+//   description: Demonstration of Sum.integersPerKey transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 46
+//   context_line: 47
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
+//     - numbers
 //     - pairs
 
-public class KvSwapExample {
+public class SumPerKeyExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
-    // Create key/value pairs
-    PCollection<KV<String, Integer>> pairs =
+    PCollection<KV<String, Integer>> input =
         pipeline.apply(
-            Create.of(KV.of("one", 1), KV.of("two", 2), KV.of("three", 3), KV.of("four", 4)));
-    // Returns KV collection with keys and values swapped: PCollection<KV<K,V>> ->
-    // PCollection<KV<V,K>>
-    PCollection<KV<Integer, String>> swap = pairs.apply(KvSwap.create());
+            Create.of(KV.of("a", 1), KV.of("a", 2), KV.of("b", 3), KV.of("b", 4), KV.of("b", 5)));
+    PCollection<KV<String, Integer>> sumPerKey = input.apply(Sum.integersPerKey());
     // [END main_section]
-    pairs.apply(ParDo.of(new LogOutput<>("PCollection element before KvSwap transform: ")));
-    swap.apply(ParDo.of(new LogOutput<>("PCollection element after KvSwap transform: ")));
+    // Log values
+    sumPerKey.apply(ParDo.of(new LogOutput<>("PCollection numbers after Sum transform: ")));
     pipeline.run();
   }
