+++ examples/java/src/main/java/org/apache/beam/examples/ValuesExample.java
@@ -15,55 +15,49 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.MapElements;
 import org.apache.beam.sdk.transforms.ParDo;
-import org.apache.beam.sdk.transforms.SimpleFunction;
+import org.apache.beam.sdk.transforms.Values;
+import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: MapElementsDemo
-//   description: Demonstration of MapElements transform usage.
+//   name: ValuesDemo
+//   description: Demonstration of Values transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 47
+//   context_line: 46
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - strings
-//     - map
+//     - pairs
 
-public class MapElementsExample {
+public class ValuesExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
-    // Create collection of lowercase string
-    PCollection<String> strings = pipeline.apply(Create.of("one", "two", "three", "four"));
-    // Transform strings to upper case
-    PCollection<String> upperCaseStrings =
-        strings.apply(
-            MapElements.via(
-                new SimpleFunction<String, String>() {
-                  @Override
-                  public String apply(String s) {
-                    return s.toUpperCase();
-                  }
-                }));
+    // Create key/value pairs
+    PCollection<KV<String, Integer>> pairs =
+        pipeline.apply(
+            Create.of(KV.of("one", 1), KV.of("two", 2), KV.of("three", 3), KV.of("four", 4)));
+    // Returns only the values of the collection: PCollection<KV<K,V>> ->
+    // PCollection<V>
+    PCollection<Integer> valuesOnly = pairs.apply(Values.create());
     // [END main_section]
-    strings.apply(ParDo.of(new LogOutput<>("PCollection element before MapElements transform: ")));
-    upperCaseStrings.apply(
-        ParDo.of(new LogOutput<>("PCollection element after MapElements transform: ")));
+    pairs.apply(ParDo.of(new LogOutput<>("PCollection element before Values.create transform: ")));
+    valuesOnly.apply(
+        ParDo.of(new LogOutput<>("PCollection element after Values.create transform: ")));
     pipeline.run();
   }
