+++ examples/java/src/main/java/org/apache/beam/examples/WithKeysExample.java
@@ -15,49 +15,53 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.examples.basic;
+package org.apache.beam.examples;
 
-import java.util.List;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.transforms.ApproximateQuantiles;
 import org.apache.beam.sdk.transforms.Create;
 import org.apache.beam.sdk.transforms.DoFn;
 import org.apache.beam.sdk.transforms.ParDo;
+import org.apache.beam.sdk.transforms.SerializableFunction;
+import org.apache.beam.sdk.transforms.WithKeys;
+import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: ApproximateQuantilesDemo
-//   description: Demonstration of ApproximateQuantiles transform usage.
+//   name: WithKeysDemo
+//   description: Demonstration of WithKeys transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 46
+//   context_line: 48
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - numbers
+//     - strings
+//     - pairs
 
-public class ApproximateQuantilesExample {
+public class WithKeysExample {
   public static void main(String[] args) {
     PipelineOptions options = PipelineOptionsFactory.create();
     Pipeline pipeline = Pipeline.create(options);
     // [START main_section]
-    // Create a collection with numbers
-    PCollection<Integer> input = pipeline.apply(Create.of(1, 1, 2, 2, 3, 4, 4, 5, 5));
-    // This will produce a collection containing 5 values: the minimum value,
-    // Quartile 1 value, Quartile 2 value (median), Quartile 3 value, and the
-    // maximum value.
-    PCollection<List<Integer>> result = input.apply(ApproximateQuantiles.globally(5));
+    PCollection<String> words = pipeline.apply(Create.of("Hello", "World", "Beam", "is", "fun"));
+    PCollection<KV<Integer, String>> lengthAndWord =
+        words.apply(
+            WithKeys.of(
+                new SerializableFunction<String, Integer>() {
+                  @Override
+                  public Integer apply(String s) {
+                    return s.length();
+                  }
+                }));
     // [END main_section]
-    result.apply(
-        ParDo.of(
-            new LogOutput<>(
-                "PCollection numbers after ApproximateQuantiles.globally transform: ")));
+    lengthAndWord.apply(
+        ParDo.of(new LogOutput<>("PCollection elements after WithKeys transform: ")));
     pipeline.run();
   }
